+++ CHANGELOG.md
@@ -793,6 +793,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Search every level
 
+## 0200 - Number of Islands
+
+- Deep First Search
+
 ## 0415 - Add strings
 
 - Full adder